refactor: move `generate` command to separate file
